requirements.txt
python-docx>=0.8.11
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.12.0
streamlit>=1.44.0
# streamlit>=1.32.0
pandas>=2.0.0
openpyxl>=3.0.0  # для работы с xlsx файлами 